+++ seq2science/cli.py
@@ -164,6 +164,7 @@ def seq2science_parser(workflows_dir="./seq2science/workflows/"):
         "--rerun-incomplete", help="Re-run all jobs the output of which is recognized as incomplete.", action="store_true"
     )
     run.add_argument("--unlock", help="Remove a lock on the working directory.", action="store_true")
+    run.add_argument("--cleanup-metadata", help="Just cleanup metadata of given list of output files (default None).", default=None, nargs="+")
     explain.add_argument("--hyperref", help="Print urls as html hyperref", action="store_true")
     # run/explain arguments
     for subparser in [run, explain]:
@@ -331,6 +332,7 @@ def _run(args, base_dir, workflows_dir, config_path):
         "printreason": args.reason,
         "keepgoing": args.keep_going,
         "unlock": args.unlock,
+        "cleanup_metadata": args.cleanup_metadata,
         "force_incomplete": args.rerun_incomplete,
     }
 
@@ -365,7 +367,7 @@ def _run(args, base_dir, workflows_dir, config_path):
     else:
         parsed_args["cores"] = 0
 
-    if parsed_args["cores"] < 2:
+    if parsed_args["cores"] < 2 and not (args.unlock or args.cleanup_metadata is not None):
         subjectively_prettier_error(core_arg, "specify at least two cores.")
 
     # when running on a cluster assume cores == nodes (just like snakemake does)
@@ -389,7 +391,7 @@ def _run(args, base_dir, workflows_dir, config_path):
         import json
         logger.info(json.dumps(parsed_args, sort_keys=True, indent=2))
 
-    if not args.skip_rerun or args.unlock:
+    if not args.skip_rerun or args.unlock or args.cleanup_metadata is not None:
         #   2. start a dryrun checking which files need to be created, and check if
         #      any params changed, which means we have to remove those files and
         #      continue from there
@@ -399,7 +401,8 @@ def _run(args, base_dir, workflows_dir, config_path):
         )
 
         with seq2science.util.CaptureStdout() as targets, seq2science.util.CaptureStderr() as errors:
-            exit_code = snakemake.snakemake(
+            exit_code = run_snakemake(
+                args.workflow.replace("-", "_"),
                 **{
                     **parsed_args,
                     **{
@@ -414,6 +417,7 @@ def _run(args, base_dir, workflows_dir, config_path):
             nl = "\n"
             logger.info(f"""Targets:\n{nl.join(sorted(targets))}\n\n""")
             logger.info(f"""Errors:\n{nl.join(sorted(errors))}\n\n""")
+
         if not exit_code:
             os._exit(1)  # noqa
 
@@ -440,7 +444,7 @@ def _run(args, base_dir, workflows_dir, config_path):
     parsed_args["config"]["no_config_log"] = True
 
     #   5. start the "real" run where jobs actually get started
-    exit_code = snakemake.snakemake(**parsed_args)
+    exit_code = run_snakemake(args.workflow.replace("-", "_"), **parsed_args)
 
     #   6. output exit code 0 for success and 1 for failure
     os._exit(0) if exit_code else os._exit(1)  # noqa
@@ -512,7 +516,7 @@ def _explain(args, base_dir, workflows_dir, config_path):
     # run snakemake (silently)
     with open(os.devnull, "w") as null:
         with contextlib.redirect_stdout(null), contextlib.redirect_stderr(null):
-            success = snakemake.snakemake(**parsed_args)
+            success = run_snakemake(args.workflow.replace("-", "_"), **parsed_args)
 
     if args.debug:
         print(f"Explain output:\n{rules_used}\n\n")
@@ -674,6 +678,31 @@ def setup_seq2science_logger(parsed_args):
         logger.get_logfile = lambda: seq2science_logfile
         logger.logfile_handler = _logging.FileHandler(seq2science_logfile)
         logger.logger.addHandler(logger.logfile_handler)
+
+
+def run_snakemake(workflow, **config):
+    try:
+        exit_code = snakemake.snakemake(**config)
+    except snakemake.exceptions.IncompleteFilesException as e:
+        exception = str(e)
+        exception = re.sub("snakemake --cleanup-metadata <filenames>",
+                           f"seq2science run {workflow} --cleanup-metadata <filenames>",
+                           exception)
+        logger.error(exception)
+        sys.exit(1)
+    except snakemake.exceptions.LockException as e:
+        exception = str(e)
+        exception = re.sub("no other Snakemake process",
+                           "no other seq2science process",
+                           exception)
+        exception = re.sub("no other instances of snakemake are",
+                           "no other instances of seq2science are",
+                           exception)
+        logger.error(exception)
+        sys.exit(1)
+    except Exception as e:
+        raise e
+    return exit_code
 
 
 def _deseq(args, base_dir):